Add mainnetClient function for consistent Ethereum mainnet RPC connection
- Introduced mainnetClient function in mainnet_client_ext_test.go and mainnet_client_test.go to streamline the connection to a public Ethereum mainnet RPC.
- Updated existing tests to utilize mainnetClient instead of hardcoded Infura URLs, enhancing test reliability and configurability.
- Ensured live-RPC tests can be skipped under `go test -short` to maintain CI compatibility.
- Removed redundant client connection code from various test files, promoting code reuse and clarity.
