+++ resolver_test.go
@@ -20,42 +20,45 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
-var client, _ = ethclient.Dial("https://mainnet.infura.io/v3/831a5442dc2e4536a9f8dee4ea1707a6")
-
 func TestResolveEmpty(t *testing.T) {
+	client := mainnetClient(t)
 	_, err := Resolve(context.Background(), client, "")
 	assert.NotNil(t, err, "Resolved empty name")
 }
 
 func TestResolveZero(t *testing.T) {
+	client := mainnetClient(t)
 	_, err := Resolve(context.Background(), client, "0")
 	assert.NotNil(t, err, "Resolved empty name")
 }
 
 func TestResolveNotPresent(t *testing.T) {
+	client := mainnetClient(t)
 	_, err := Resolve(context.Background(), client, "sirnotappearinginthisregistry.eth")
 	require.NotNil(t, err, "Resolved name that does not exist")
 	assert.Equal(t, "unregistered name", err.Error(), "Unexpected error")
 }
 
 // func TestResolveNoResolver(t *testing.T) {
+// 	client := mainnetClient(t)
 // 	_, err := Resolve(context.Background(), client, "noresolver.eth")
 // 	require.NotNil(t, err, "Resolved name without a resolver")
 // 	assert.Equal(t, "no resolver", err.Error(), "Unexpected error")
 // }
 
 func TestResolveBadResolver(t *testing.T) {
+	client := mainnetClient(t)
 	_, err := Resolve(context.Background(), client, "resolvestozero.eth")
 	require.NotNil(t, err, "Resolved name with a bad resolver")
 	assert.Equal(t, "no address", err.Error(), "Unexpected error")
 }
 
 func TestResolveTestEnsTest(t *testing.T) {
+	client := mainnetClient(t)
 	expected := "ed96dd3be847b387217ef9de5b20d8392a6cdf40"
 	actual, err := Resolve(context.Background(), client, "test.enstest.eth")
 	require.Nil(t, err, "Error resolving name")
@@ -63,6 +66,7 @@ func TestResolveTestEnsTest(t *testing.T) {
 }
 
 func TestResolveResolverEth(t *testing.T) {
+	client := mainnetClient(t)
 	expected := "231b0ee14048e9dccd1d247744d114a4eb5e8e63"
 	actual, err := Resolve(context.Background(), client, "resolver.eth")
 	require.Nil(t, err, "Error resolving name")
@@ -70,6 +74,7 @@ func TestResolveResolverEth(t *testing.T) {
 }
 
 func TestResolveAddress(t *testing.T) {
+	client := mainnetClient(t)
 	expected := "b8c2c29ee19d8307cb7255e1cd9cbde883a267d5"
 	actual, err := Resolve(context.Background(), client, "0xb8c2C29ee19D8307cb7255e1Cd9CbDE883A267d5")
 	require.Nil(t, err, "Error resolving address")
@@ -77,6 +82,7 @@ func TestResolveAddress(t *testing.T) {
 }
 
 func TestResolveShortAddress(t *testing.T) {
+	client := mainnetClient(t)
 	expected := "0000000000000000000000000000000000000001"
 	actual, err := Resolve(context.Background(), client, "0x1")
 	require.Nil(t, err, "Error resolving address")
@@ -84,11 +90,13 @@ func TestResolveShortAddress(t *testing.T) {
 }
 
 func TestResolveHexString(t *testing.T) {
+	client := mainnetClient(t)
 	_, err := Resolve(context.Background(), client, "0xe32c6d1a964749b6de2130e20daed821a45b9e7261118801ff5319d0ffc6b54a")
 	assert.NotNil(t, err, "Resolved too-long hex string")
 }
 
 func TestReverseResolveTestEnsTest(t *testing.T) {
+	client := mainnetClient(t)
 	expected := "nick.eth"
 	address := common.HexToAddress("b8c2C29ee19D8307cb7255e1Cd9CbDE883A267d5")
 	actual, err := ReverseResolve(context.Background(), client, address)
