+++ reverseresolver_test.go
@@ -19,7 +19,6 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/require"
 	ens "github.com/wealdtech/go-ens/v4"
 )
@@ -51,8 +50,7 @@ func TestReverseResolve(t *testing.T) {
 		},
 	}
 
-	client, err := ethclient.Dial("https://mainnet.infura.io/v3/831a5442dc2e4536a9f8dee4ea1707a6")
-	require.NoError(t, err)
+	client := mainnetClient(t)
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
