+++ tokenid_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"testing"
 
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/stretchr/testify/require"
 )
 
@@ -36,8 +35,7 @@ func TestDeriveTokenId(t *testing.T) {
 			err:      "empty domain",
 		},
 	}
-	client, err := ethclient.Dial("https://mainnet.infura.io/v3/831a5442dc2e4536a9f8dee4ea1707a6")
-	require.NoError(t, err)
+	client := mainnetClient(t)
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			actual, err := DeriveTokenID(context.Background(), client, test.input)
